% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_wrapper.R
\name{annotate_DESeqOutput}
\alias{annotate_DESeqOutput}
\title{annotate DESeq2 Output file}
\usage{
annotate_DESeqOutput(DESeqOutput, Output, remote = TRUE, genome = "hg38",
  map_file)
}
\arguments{
\item{DESeqOutput}{Tab-seperated DESeq2 output file}

\item{Output}{Annotated output file}

\item{remote}{Whether use biomart to annotate file}

\item{genome}{When remote = TRUE, which genome to use? (available = "hg38","mm10","dm6")}

\item{map_file}{If remote = FALSE, provide a map file (with ENS id and Gene id in column 1 and 2) respectively}
}
\value{
annotated output file
}
\description{
annotate DESeq2 Output file
}
\examples{
annotate_DESeqOutput(DESeqOutput = "test.out", Output = "test_annotated.out", remote = TRUE, genome = "hg38")

}

