% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera_wrapper.R
\name{Camera_plotbubble}
\alias{Camera_plotbubble}
\title{Make a bubble plot for CAMERA output}
\usage{
Camera_plotbubble(CAMERAoutput, outfileName, top = 20, title = NULL)
}
\arguments{
\item{CAMERAoutput}{A tab-seperated file with CAMERA results.}

\item{outfileName}{Output pdf file name.}

\item{top}{Number of top gene-sets to plot.}

\item{title}{Title of the plot}
}
\value{
A pdf file with bubble plot
}
\description{
Make a bubble plot for CAMERA output
}
\examples{

Camera_plotbubble(CAMERAoutput, outfileName, top = 20, title = NULL)

}

