% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffExp_wrapper.R
\name{EdgeR_wrapper}
\alias{EdgeR_wrapper}
\title{A Wrapper for EdgeR over featurecounts output}
\usage{
EdgeR_wrapper(fcountOutput, numReplicates = 4, fdr = 0.01,
  Output = "edger_output.tsv", col_order = 1, heatmap_topN = 20,
  pdfReport = "edger_report.pdf")
}
\arguments{
\item{fcountOutput}{featurecounts output (with control and test columns)}

\item{numReplicates}{Number of replicates (could be an integer if the number is same for control and test,
or a vector with number of replicates for control and for test seperately)}

\item{fdr}{fdr Cutoff}

\item{Output}{Output tab seperated file}

\item{col_order}{How are the columns arranged? (1 = control->test, 2 = test->control)}

\item{heatmap_topN}{Number of top DE genes to plot in a heatmap (will make rlog trasformed and clustered heatmap)
type "all" for all DE genes.}

\item{pdfReport}{A report of processing}
}
\value{
Output file and Pdf Report of DESeq2 analysis
}
\description{
A Wrapper for EdgeR over featurecounts output
}
\examples{
fc <- system.file("extdata", "fcount_mouse.tsv", package="vivlib")
EdgeR_wrapper(fcountOutput = fc,numReplicates = 3, fdr = 0.01,
Output = "edger_output.tsv", pdfReport = "edger_report.pdf")
}

