% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_annotation}
\alias{fetch_annotation}
\title{Fetch annotation from ENSEMBL ID to gene ID from Biomart}
\usage{
fetch_annotation(genome)
}
\arguments{
\item{genome}{genome name ("hg38","mm10","dm6")}
}
\value{
annotation
}
\description{
Fetch annotation from ENSEMBL ID to gene ID from Biomart
}

