% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIntrons_byGenes.R
\name{getIntrons_byGenes}
\alias{getIntrons_byGenes}
\title{Get introns by genes}
\usage{
getIntrons_byGenes(txdb, seqout = "DESeq_outputs/XR1_annotated.out",
  padjFilter = 0.05, genome_mart = "hsapiens_gene_ensembl", outfile)
}
\arguments{
\item{txdb}{A TxDb object}

\item{seqout}{The annotated DESeq Output}

\item{padjFilter}{p-value cutoff}

\item{genome_mart}{the name of mart object to annotate the input with (eg : hsapiens_gene_ensembl)}

\item{outfile}{output file with introns, to write back}
}
\value{
annotated DESeq output with introns for each gene
}
\description{
Get introns by genes
}
\examples{
library(TxDb.Hsapiens.UCSC.hg38.knownGene)

getIntrons_byGenes(txdb = TxDb.Hsapiens.UCSC.hg38.knownGene, seqout = "DESeq_outputs/XR1_annotated.out",
padjFilter = 0.05, genome_mart = "hsapiens_gene_ensembl", outfile = "annotated_introns.out")

}

