% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean_bysample}
\alias{mean_bysample}
\title{A function to get normalized mean of replicates by sample}
\usage{
mean_bysample(name, df)
}
\arguments{
\item{name}{name of sample (from the given df)}

\item{df}{A data frame of counts}
}
\value{
data frame
}
\description{
A function to get normalized mean of replicates by sample
}

