% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBarCodes.R
\name{plotBarCodes}
\alias{plotBarCodes}
\title{Perform ROAST and make BarCodePlot using a voom transformed output and a microarry output}
\usage{
plotBarCodes(GSEfile, ourVoomFile, batchAnalyse = TRUE,
  VoomInputName = NULL,
  humanMouseNameMap = "/data/akhtar/bhardwaj/my_annotations/Human_mouse_Gene_orthologous.txt",
  dfCutoff = "pvalue", logFoldCh = 0, padj = 0.05,
  outFolder = "/data/akhtar/bhardwaj/2015_OtherAnalysis/Bilal_Oncogene_paper/BarCodePlots/MPC5")
}
\arguments{
\item{GSEfile}{A microarry data file. Filename should start from 'GSE'.
The columns should have 'gname' and 'SPOT_ID'.}

\item{ourVoomFile}{Voom transformed output from \code{\link{makeVoomInput}}}

\item{batchAnalyse}{If there are multiple microarry files in the folder, give the folder path.}

\item{VoomInputName}{Provide a name for our Voom-transformed input data.}

\item{humanMouseNameMap}{A tab-seperated file with column ("Sym_Human") which provides alternative
human symbol for our mouse gene IDs}

\item{dfCutoff}{Which cutoff to use for filtering genes from GSEfile.
Default is by pvalue. Otherwise a log Fold Change cutoff can be used.}

\item{logFoldCh}{logFoldChange cutoff for filtering GSE file (if used)}

\item{padj}{adjusted p-value cutoff for filtering GSE file (if used)}

\item{outFolder}{Folder to write the output pdf file.}
}
\value{
Pdf file with barcode plots
}
\description{
This function was originally written for comparing a  voom-transformed mouse input
and a microarry (affy) input from human or mouse. That's why the option to provide a Human-Mouse ID map.
But if you don't want that, you can simply use it to compare any set of microarry outputs to any voom-transformed input.
}
\examples{
\dontrun{
     plotBarCodes(GSEfile,ourVoomFile)
}

}

