% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEgene_clustering_Intersections.R
\name{plotDEgeneOverlap}
\alias{plotDEgeneOverlap}
\title{plot intersections of list of DE genes from multiple DESeq2 outputs}
\usage{
plotDEgeneOverlap(DEoutList, sampleNames, FDRcutoff = 0.05, outFile = NULL,
  returnData = FALSE, ...)
}
\arguments{
\item{DEoutList}{A vector, with names of DESeq2 output files to use.}

\item{sampleNames}{Name of samples corresponding to the DESeq output file list.}

\item{FDRcutoff}{FDR cutoff to select DE genes from the list}

\item{outFile}{Name of output pdf file. If NULL, prints output on the screen}

\item{returnData}{Return the UpSet-formatted data}

\item{...}{Additional parameters passed to the function `upSet`}
}
\value{
A plot of Intersection of gene sets.
}
\description{
plot intersections of list of DE genes from multiple DESeq2 outputs
}
\examples{
\dontrun{
plotDEgeneOverlap(DEoutList, sampleNames, FDRcutoff = 0.05, outFile = NULL)
}
}

