% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeatureCount.R
\name{plotFeatureCounts}
\alias{plotFeatureCounts}
\title{plot the output of featureCounts summary}
\usage{
plotFeatureCounts(summaryFile, CutFromHeader, outFile)
}
\arguments{
\item{summaryFile}{featureCounts output.summary file}

\item{CutFromHeader}{unnecessory text to remove from the header}
}
\value{
A plot of featurecounts summary in a pdf file
}
\description{
plot the output of featureCounts summary
}
\examples{
plotFeatureCounts(test.summary,"/long/path/to/cut")

}

