% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_wrappers.R
\name{plotPathway}
\alias{plotPathway}
\title{Plot DESeq2 output on a selected KEGG pathway map}
\usage{
plotPathway(DESeqOutput, fdr = 0.05, pathway_IDs = c("05200", "04110"),
  genome = "hg38")
}
\arguments{
\item{DESeqOutput}{A tab-seperated DESeq2 output file}

\item{fdr}{FDR cutoff for Diff-Exp genes.}

\item{pathway_IDs}{KEGG ids for the pathways to plot}

\item{genome}{Select from hg38, mm10, dm6. (although it's agnostic to genome versions in general).}
}
\value{
KEGG plot
}
\description{
Plot DESeq2 output on a selected KEGG pathway map
}
\examples{


}

