% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spia_wrapper.R
\name{spia_plotBubble}
\alias{spia_plotBubble}
\title{Make a bubblePlot for SPIA pathway output}
\usage{
spia_plotBubble(SPIAout, outfileName = NULL, top = 20, plotType = 1,
  title = NULL)
}
\arguments{
\item{SPIAout}{output from spia_wrapper}

\item{outfileName}{output pdf file name for plot}

\item{top}{How many top pathways to plot (by pGFdr value)}

\item{plotType}{Which kind of bubble plot to make. options are :
1 (activated and inactivated pathways together) or
2 (activated and inactivated pathways split by a horizontal line).
Note that in type 2 the p values appear as negative for inactivated pathways.}

\item{title}{Title of the plot}
}
\value{
plot A bubbleplot in pdf format
}
\description{
Make a bubblePlot for SPIA pathway output
}
\examples{
\dontrun{
deout <- system.file("extdata", "edgeR_output_annotated.tsv", package="vivlib")
spia_wrapper(DEoutput = deout, organism = "mmu", outFile = "edgeR_spia.txt")
spia_plotBubble(SPIAout = "edgeR_spia.txt", outfileName = "SPIAplots.out", top = 20, title = "SPIA plot")
}
}

