% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spia_wrapper.R
\name{spia_wrapper}
\alias{spia_wrapper}
\title{A wrapper over SPIA pathway perturbation analysis for mouse and human}
\usage{
spia_wrapper(DEoutput, organism = "mmu", padjCutoff = 0.05, outFile,
  outPlot = NULL)
}
\arguments{
\item{DEoutput}{A tab-seperated edgeR/DESeq2 output file, using \code{\link{EdgeR_wrapper}} or
\code{\link{DESeq_wrapper}}}

\item{organism}{select from "mmu" (mouse) or "hsa" (human)}

\item{padjCutoff}{FDR cutoff for the genes to include in the list}

\item{outFile}{output file to write affected pathways}

\item{outPlot}{output SPIA two-way evidence plot + output summary (pdf)}
}
\value{
a txt file and a pdf file
}
\description{
A wrapper over SPIA pathway perturbation analysis for mouse and human
}
\examples{
\dontrun{
deout <- system.file("extdata", "edgeR_output_annotated.tsv", package="vivlib")
spia_wrapper <- function(DEoutput = deout, organism = "mmu", outFile = "edgeR_spia.txt")
}
}

