% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEgene_clustering_Intersections.R
\name{clusterDEgenes}
\alias{clusterDEgenes}
\title{Cluster DE genes by fold change from multiple files}
\usage{
clusterDEgenes(DEoutList, sampleNames, FDRcutoff = 0.05,
  method = "correlation", cut_cluster = NA, row_annotation = NA,
  outFile_prefix = NULL)
}
\arguments{
\item{DEoutList}{A vector, with names of DESeq2 output files to use.}

\item{sampleNames}{Name of samples corresponding to the DESeq output file list.}

\item{FDRcutoff}{FDR cutoff to select DE genes from the list}

\item{method}{Which clustering method to use. Either "correlation" or "biclustering"}

\item{cut_cluster}{A number to which the cluster dendrogram will be cut into (NA means do not cut clusters)}

\item{row_annotation}{A data frame for the annotation of genes, with rownames corresponding to the Row.names column of the DESeq2 output. The columns can have annotations like chromosome, gene type etc..}

\item{outFile_prefix}{A prefix for output files.}
}
\value{
A pdf with clustered heatmap, an .Rdata file with the hclust objects and the genes sorted by clustered output, and text file with genes divided by clusters (if cut_clusters is selected).
}
\description{
Cluster DE genes by fold change from multiple files
}
\examples{

clusterDEgenes(DEoutList, sampleNames, FDRcutoff = 0.05, method = "correlation")

}

