% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_wrapper.R
\name{DESeq_wrapper}
\alias{DESeq_wrapper}
\title{A Wrapper fr DESeq2 over featurecounts output}
\usage{
DESeq_wrapper(fcountOutput, numReplicates = 4, fdr = 0.01,
  Output = "deseq_output.tab", pdfReport = "deseq_report.pdf")
}
\arguments{
\item{fcountOutput}{featurecounts output (with control and test columns)}

\item{numReplicates}{Number of replicates}

\item{fdr}{fdr Cutoff}

\item{Output}{Output tab seperated file}

\item{pdfReport}{A report of processing}
}
\value{
Output file and Pdf Report of DESeq2 analysis
}
\description{
A Wrapper fr DESeq2 over featurecounts output
}
\examples{
DESeq_wrapper(fcountOutput = "test.out",numReplicates = 4, fdr = 0.01, Output = "deseq_output.tab",
             pdfReport = "deseq_report.pdf")
}

