% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureCounts_downstream.R
\name{fcount_meantransform}
\alias{fcount_meantransform}
\title{Transform a featurecount output to a data frame of library-normalized mean counts}
\usage{
fcount_meantransform(fcountOutput, samplenames, filterByCount = NULL,
  boxplot = TRUE)
}
\arguments{
\item{samplenames}{The name of samples (exclusing replicate ID etc). Replicates of the samples
will be grouped together for calculating mean..}

\item{filterByCount}{A value, if given it will first filter genes by raw mean expression <= given threshold.}

\item{boxplot}{Logical. Draw a boxplot or normalized counts?}

\item{fcoutOutput}{Featurecount output in a data frame}
}
\value{
A data frame of library-normalized mean counts
}
\description{
Transform a featurecount output to a data frame of library-normalized mean counts
}
\examples{
fc <- system.file("extdata", "fcount_mouse.out", package="vivlib")
samples <- rep(c("cnt","KD"), each = 3)
fcount_meantransform(fcoutOutput = fc, samplenames = samples, filterByCount = 1000, boxplot = TRUE)

}

