% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBarCodes.R
\name{makeVoomInput}
\alias{makeVoomInput}
\title{Make Voom transformed input for BarCodePlot and ROAST test}
\usage{
makeVoomInput(fcountOutput, design, genome = "mm10", plotFile = NULL)
}
\arguments{
\item{fcountOutput}{A featurecounts output file}

\item{design}{A data frame containing design information. rownames should correspond to columns in count file and
column could be control/test or any set of factors.}

\item{plotFile}{File name to output filtering plots (.pdf).}
}
\value{
A plot (as pdf) and a voom transformed output as list.
}
\description{
Make Voom transformed input for BarCodePlot and ROAST test
}
\examples{
fc <- system.file("extdata", "fcount_mouse.out", package="vivlib")
design <- data.frame(rownames = c(paste0("cnt_", 1:3), paste0("KD_",1:3)),
                     condition = rep(c("cnt","KD"), each = 3) )
makeVoomInput(fcountOutput = fc, design = design, plotFile = "test.pdf")

}

