% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeatureCount.R
\name{plotFeatureCounts}
\alias{plotFeatureCounts}
\title{plot the output of featureCounts summary}
\usage{
plotFeatureCounts(summaryFile, CutFromHeader)
}
\arguments{
\item{summaryFile}{featureCounts output.summary}

\item{CutFromHeader}{Enter the long path that's present in featurecounts output to cut}
}
\value{
a plot
}
\description{
plot the output of featureCounts summary
}
\examples{
plotFeatureCounts(test.summary,"/long/path/to/cut")

}

