% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_wrappers.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot heatmap of raw counts for top DEgenes using DESeq2 output}
\usage{
plotHeatmap(DESeqOutput, fdr = 0.05, fcountOutput, sampleNames,
  topNgenes = 100, clusterbyCorr = FALSE, useGeneNames = TRUE,
  markGenes = NULL, outFile = NULL)
}
\arguments{
\item{DESeqOutput}{A tab-seperated DESeq2 output file}

\item{fdr}{FDR cutoff for DE genes}

\item{fcountOutput}{Featurecounts output (for raw counts)}

\item{sampleNames}{samplenames for heatmap column label (must be the same order as in featurecounts file)}

\item{topNgenes}{How many genes to plot. Type NULL for all genes (might take long time)}

\item{clusterbyCorr}{Whether to cluster row/columns by correlations. Default is eucledian distances.}

\item{useGeneNames}{Use gene names to plot instead of default geneIDs. only works if the input
has been annotated by \code{\link{annotate_DESeqOutput}} function.}

\item{markGenes}{Genes to mark in bold on heatmap.}

\item{outFile}{File name to save the output. NULL prints heatmap on screen.}
}
\value{
A heatmap.
}
\description{
Plot heatmap of raw counts for top DEgenes using DESeq2 output
}
\examples{


}

