% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_wrappers.R
\name{plotPathway}
\alias{plotPathway}
\title{Plot DE output on a selected KEGG pathway map}
\usage{
plotPathway(DEoutput, fdr = 0.05, pathway_IDs = c("05200", "04110"),
  genome = "hg38")
}
\arguments{
\item{DEoutput}{A tab-seperated edgeR/DESeq2 output file, using \code{\link{EdgeR_wrapper}} or
\code{\link{DESeq_wrapper}}}

\item{fdr}{FDR cutoff for Diff-Exp genes.}

\item{pathway_IDs}{KEGG ids for the pathways to plot}

\item{genome}{Select from hg38, mm10, dm6. (although it's agnostic to genome versions in general).}
}
\value{
KEGG plot
}
\description{
Plot DE output on a selected KEGG pathway map
}
\examples{
deout <- system.file("extdata", "edgeR_output_annotated.tsv", package="vivlib")
plotPathway(DEoutput = deout, fdr = 0.05, pathway_IDs = "05200", genome = "mm10")

}

