% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera_wrapper.R
\name{runCamera}
\alias{runCamera}
\title{A wrapper over CAMERA test.}
\usage{
runCamera(counts, design, bmGeneNames, moduleFile = "msigdb.v5.0.symbols.gmt",
  outfileName)
}
\arguments{
\item{counts}{A tab-seperated file containing gene names followed by counts}

\item{design}{A tab-seperated file containing design information (colnames, condition).
colnames should correspond to columns in count file and condition could be
control/test or any set of factors.}

\item{bmGeneNames}{Optionally provide alternative gene symbols downloaded from biomart as a
tab-seperated file. The columns should be ("ensembl_gene_id","external_gene_id")}

\item{moduleFile}{A file with modules to test. Download from
\code{\link{http://software.broadinstitute.org/gsea/msigdb/index.jsp}}}

\item{outfileName}{Output pdf file name.}

\item{name}{File name to output filtering plots.}
}
\value{
CAMERA result object.
}
\description{
A wrapper over CAMERA test.
}
\examples{
runCamera(counts,design,bmGeneNames,name="myVoomInput",moduleFile="msigdb.v5.0.symbols.gmt")

}

