% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera_wrapper.R
\name{runCamera}
\alias{runCamera}
\title{A wrapper over CAMERA test.}
\usage{
runCamera(fcountOutput, design, bmGeneNames,
  moduleFile = "msigdb.v5.0.symbols.gmt", outfileName)
}
\arguments{
\item{fcountOutput}{A tab-seperated file containing gene names followed by fcountOutput}

\item{design}{A tab-seperated file containing design information (colnames, condition).
colnames should correspond to columns in count file and condition could be
control/test or any set of factors.}

\item{bmGeneNames}{Optionally provide alternative gene symbols downloaded from biomart as a
tab-seperated file. The columns should be ("ensembl_gene_id","external_gene_id")}

\item{moduleFile}{A file with modules to test. Download from
\url{http://software.broadinstitute.org/gsea/msigdb/index.jsp}}

\item{outfileName}{Output pdf file name.}

\item{name}{File name to output filtering plots.}
}
\value{
CAMERA result object.
}
\description{
A wrapper over CAMERA test.
}
\examples{

\dontrun{
fc <- system.file("extdata", "fcount_mouse.out", package="vivlib")
design <- data.frame(rownames = c(paste0("cnt_", 1:3), paste0("KD_",1:3)),
                     condition = rep(c("cnt","KD"), each = 3) )

runCamera(fcountOutput,design,bmGeneNames,name="myVoomInput",outfileName = "test")
}


}

