% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spia_wrapper.R
\name{spia_plotBubble}
\alias{spia_plotBubble}
\title{Make a bubblePlot for SPIA pathway output}
\usage{
spia_plotBubble(SPIAout, outfileName, top = 20, title = NULL)
}
\arguments{
\item{SPIAout}{output from spia_wrapper}

\item{outfileName}{output pdf file name for plot}

\item{top}{How many top pathways to plot (by pGFdr value)}

\item{title}{Title of the plot}
}
\value{
plot A bubbleplot in pdf format
}
\description{
Make a bubblePlot for SPIA pathway output
}
\examples{
spia_plotBubble(spia_wrapper_output, outfileName = "test.out, top = 20, title = "test plot)

}

