% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spia_wrapper.R
\name{spia_wrapper}
\alias{spia_wrapper}
\title{A wrapper over SPIA pathway analysis for mouse and human}
\usage{
spia_wrapper(DESeqOut, organism = "mmu", padjCutoff, outFile, outPlot)
}
\arguments{
\item{DESeqOut}{Tab-seperated DESeq2 output}

\item{organism}{select from "mmu" (mouse) or "hsa" (human)}

\item{padjCutoff}{FDR cutoff for the genes to include in the list}

\item{outFile}{output file to write affected pathways}

\item{outPlot}{output SPIA two-way evidence plot + output summary (pdf)}
}
\value{
a txt file and a pdf file
}
\description{
A wrapper over SPIA pathway analysis for mouse and human
}
\examples{


}

